refactor: 边框8 tsx 改造
