+++ packages/datav-vue3/components/BorderBox8/index.ts
@@ -1,5 +1,5 @@
 import type { App, Plugin } from 'vue'
-import BorderBox8 from './src/index.vue'
+import BorderBox8 from './src/BorderBox8'
 
 export const BorderBox8Plugin: Plugin = {
   install(app: App) {
