+++ packages/datav-vue3/components/BorderBox8/src/index.less
@@ -1,0 +1,19 @@
+.dv-border-box-8 {
+  position: relative;
+  width: 100%;
+  height: 100%;
+
+  .dv-border-svg-container {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    left: 0px;
+    top: 0px;
+  }
+
+  .border-box-content {
+    position: relative;
+    width: 100%;
+    height: 100%;
+  }
+}
